split the rpc utils
