+++ rpent/utils/rpc/client_utils.py
@@ -1,0 +1,77 @@
+"""Client utility functions: endpoint parsing, server discovery, and readiness polling."""
+
+from __future__ import annotations
+
+import time
+from typing import TYPE_CHECKING
+
+from rpent.utils.logging import get_logger
+
+if TYPE_CHECKING:
+    from rpent.utils.daemon import ProcessDaemon
+    from rpent.utils.rpc.rpc import RpcClient
+
+logger = get_logger("client_utils")
+
+
+def parse_endpoint(endpoint: str) -> tuple[str, str, int]:
+    """Parse ``[protocol://]host:port`` into ``(protocol, host, port)``.
+
+    Protocol defaults to ``http`` when the prefix is omitted.
+    """
+    if "://" in endpoint:
+        protocol, _, rest = endpoint.partition("://")
+    else:
+        protocol, rest = "http", endpoint
+    host, _, port = rest.partition(":")
+    if not host or not port:
+        raise ValueError(f"endpoint must be [protocol://]host:port, got {endpoint!r}")
+    return protocol, host, int(port)
+
+
+def make_rpc_client(endpoint: str) -> "RpcClient":
+    """Build an HTTP or socket client for ``endpoint``."""
+    from rpent.utils.rpc.http_rpc import HttpRpcClient
+    from rpent.utils.rpc.socket_rpc import SocketRpcClient
+
+    protocol, host, port = parse_endpoint(endpoint)
+    if protocol == "http":
+        return HttpRpcClient(f"http://{host}:{port}")
+    if protocol == "socket":
+        return SocketRpcClient(host, port)
+    raise ValueError(f"endpoint protocol must be http or socket, got {protocol!r}")
+
+
+def wait_for_ready(
+    client: "RpcClient",
+    *,
+    timeout_s: float = 300.0,
+    poll_interval_s: float = 0.5,
+    daemon: "ProcessDaemon | None" = None,
+) -> None:
+    """Poll ``client.call("healthz")`` until it succeeds or ``timeout_s`` elapses.
+
+    If ``daemon`` is given and its subprocess exits before becoming ready,
+    fail fast with ``RuntimeError`` (carrying the exit code) instead of
+    blocking until ``timeout_s``.
+    """
+    deadline = time.time() + timeout_s
+    last_err: Exception | None = None
+    while time.time() < deadline:
+        if daemon is not None:
+            rc = daemon.poll()
+            if rc is not None:
+                detail = last_err if last_err is not None else "no healthz attempt yet"
+                raise RuntimeError(
+                    f"{daemon.name} exited with code {rc} before becoming "
+                    f"ready; check its log. last healthz error: {detail}"
+                )
+        try:
+            client.call("healthz", timeout_s=1.0)
+            return
+        except Exception as exc:
+            last_err = exc
+            time.sleep(poll_interval_s)
+    raise TimeoutError(
+        f"server did not become ready within {timeout_s:.0f}s: {last_err}"
+    )
