+++ rpent/utils/rpc/socket_rpc.py
@@ -31,7 +31,10 @@ import socketserver
 import struct
 from typing import Any, Callable
 
-from rpent.utils.rpc import check_response, make_error_response
+from rpent.utils.logging import get_logger
+from rpent.utils.rpc.rpc import RpcClient, check_response, make_error_response
+
+logger = get_logger("socket_rpc")
 
 DEFAULT_CONNECT_TIMEOUT_S = 10.0
 DEFAULT_REQUEST_TIMEOUT_S = 30.0
@@ -62,7 +65,7 @@ def _write_frame(writer, obj: Any) -> None:
     writer.flush()
 
 
-class SocketRpcClient:
+class SocketRpcClient(RpcClient):
     """One-request-per-connection pickle-framed RPC client."""
 
     def __init__(
@@ -106,7 +109,8 @@ class _RequestHandler(socketserver.StreamRequestHandler):
     def handle(self) -> None:
         try:
             payload = _read_frame(self.rfile)
-        except Exception:
+        except Exception as exc:
+            logger.debug("rpc read failed: %s", exc)
             return
         try:
             method = payload["method"]
@@ -118,8 +122,8 @@ class _RequestHandler(socketserver.StreamRequestHandler):
             response = make_error_response(exc)
         try:
             _write_frame(self.wfile, response)
-        except Exception:
-            pass
+        except Exception as exc:
+            logger.debug("rpc write failed: %s", exc)
 
 
 class SocketRpcServer(socketserver.ThreadingTCPServer):
